Improve scripts and typing

+++ scripts/down-detached.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-kill -9 $(cat ./serverless.pid) && rm ./serverless.pid
+if [ -f "./serverless.pid" ]; then
+  kill -9 $(cat ./serverless.pid) && rm ./serverless.pid
+fi
